backend/src/gateway/requirements.txt
astroid==2.9.3
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
Flask==2.0.2
Flask-PyMongo==2.3.0
idna==3.3
isort==5.10.1
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
lazy-object-proxy==1.7.1
MarkupSafe==2.0.1
mccabe==0.6.1
parso==0.8.3
pika==1.2.0
platformdirs==2.5.0
pylint==2.12.2
pymongo==4.0.1
requests==2.27.1
toml==0.10.2
urllib3==1.26.8
Werkzeug==2.0.3
wrapt==1.13.3
psycopg2-binary